Implement RegisteredMetricsFilter (#15)
This implements a filter that makes it possible to define a set of
metrics in an external file and enforce that only those metrics are
emitted. This is helpful to reduce typos and improve documentation of
metrics.

This takes a structure rather than a path to an external file. In this
way, it makes it generic allowing users to define metrics in whatever
way they find most convenieng. e.g. JSON, YAML, Python, etc.