Rework LoggingBackend and timestamps (#11)
This adds a new "timestamp_mode" which changes the behavior of the
timestamp in the logged output.

* "utc": inserts a UTC timestamp
* "local": inserts a local timezone timestamp
* anything: doesn't insert a timestamp at all

In the cases where I'm using Markus in a local dev environment for
service development, the logging framework is already configured with a
timestamp, so having Markus' LoggingBackend also include a timestamp is
redundant and not wildly helpful. Thus I wanted to include a "no
timestamp" option.

For when the user wants a timestamp, it seemed prudent to support both
UTC and local timezone timestamps.

Also, timestamps are now emitted in isoformat.